2 minor bugs fixed jgrapht\jgrapht-core\src\main\java\org\jgrapht\traverse\ClosestFirstIterator.java -> L179 jgrapht-unimi-dsi/src/main/java/org/jgrapht/webgraph/ImmutableDirectedBigGraphAdapter.java -> L136

+++ jgrapht-unimi-dsi/src/main/java/org/jgrapht/webgraph/ImmutableDirectedBigGraphAdapter.java
@@ -133,7 +133,7 @@ public class ImmutableDirectedBigGraphAdapter
     @Override
     public int degreeOf(final Long vertex)
     {
-        final long d = inDegreeOf(vertex) + outDegreeOf(vertex);
+        final long d =(long) (inDegreeOf(vertex) + outDegreeOf(vertex));
         if (d > Integer.MAX_VALUE)
             throw new ArithmeticException();
         return (int) d;
